backend/requirements.txt
fastapi
uvicorn[standard]
numpy==1.26.4
python-dotenv
pydantic
requests